requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==3.0.0
gunicorn==21.2.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database & caching
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1

# Trading & market data
alpaca-py==0.13.3
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.33

# Scheduling & async
apscheduler==3.10.4
celery==5.3.4

# LLM & AI
openai==1.3.8
httpx==0.25.2
asyncio  # For MCP client

# Analysis & signals
ta==0.11.0
scikit-learn==1.3.2
scipy==1.11.4

# Notifications
slack-sdk==3.23.0
email-validator==2.1.0.post1

# Utilities
python-json-logger==2.0.7
pytz==2023.3.post1
click==8.1.7
PyYAML==6.0.1
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
minio==7.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
ruff==0.1.8